Refactor code
1. re-organize source code folder structure
2. refactor out helpers to utils module
3. refactor code to differentiate install vs upgrade scenario and treat
them differently for notification purposes.
