+++ Extensions/azure-node-essentials/src/codegen/codegen.js
