+++ Extensions/azure-node-essentials/src/commands/azure-node.template-deploy.js
