+++ Extensions/azure-node-essentials/src/snippets/snippets.json
